refactor: Changes runStorageT to run in Prerender
Now that hydration works in reflex-dom, we should probably make
runStorageT be more friendly to running the smallest possible prerender
slices so that users can render as much of the DOM as they can that
doesn't depend on the storage values.

This change is a little bogus, because there is no way for the user to
tell that they are getting nothing because switchover hasn't happened or
because it actually isn't in local storage. I'm not sure whether this is
a big deal or not, but it feels like it probably will cause an issue.

We probably need to fix this, but I figured that I'd put this change in
to start the conversation first and then test it out in realworld.